Rename tradequote file to tradeQuoter
