+++ src/api/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './tradequote';
+export * from './tradeQuoter';
 export * from './coingecko';
 export * from './gasOracle';
 export * from './zeroex';
